Fixed a minor issue introduced during cleaning (#12)
This issue was introduced during cleaning the Python script for linting compliance. The original code will take the key from previous round of parsing, which will trip on the linting error. This change will set it to the global variable of this class instance to avoid linting violation or functional discontinuities.

Signed-off-by: Kun Qin <kun.qin@microsoft.com>