SetupDataPkg: ConfApp: Conditionally blocking the configuration updates
This change will block the configuration updates based on the system
operation mode. When blocked, the configuration app will render the
disabled options in dark grey with a notification that these options are
disabled.

The configuration dumping is still available.

Unit test is also added to verify the options are properly blocked.