Updated conf app to selectively signal ready to boot at the entry point (#45)
Current design will allow users to update secure boot when booting directly to the configuration app as the boot menu, which might be of concern on non-client systems.

The change is to let the application signal a ready to boot event at the entrypoint when system is not in MFG mode. This will enforce the entire system to exit the security boundry, if not already exited, even on direct booting to this application.