Added binary CSV column (#76)
## Description

Added a binary column to the CSV file format to show the binary
representation of values.

This is added to aid users who need to manually update knobs by directly
editing UEFI variables. The binary value shows the correct
representation for a given value.

The value is written by the tools, but is ignored when read, similar to
the Comment/Help column. The Help column was renamed to Comment to make
clear it can be used for other notes in saved CSVs.

For each item, place an "x" in between `[` and `]` if true. Example:
`[x]`.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Manually verified

## Integration Instructions