Allow arbitrary SVD values to be set through ConfApp (#77)
## Description

Currently, mu_feature_config is in a transition to an XML based backend.
The current YAML based ConfApp expects SVDs with the IDs
`"Device.ConfigData.*"`. However, the XML has no such restriction. In
the case of the XML, we want to allow arbitrary SVD provided variables
to be set, as this option is only provided in the manufacturing mode
state (used often for bringup and debug).

This change leaves the YAML processing in place until more XML pieces
are set to remove the remaining YAML pieces.

For each item, place an "x" in between `[` and `]` if true. Example:
`[x]`.
_(you can also check items in the GitHub UI)_

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Tested with unit tests and in QemuQ35 by booting to the ConfApp in
manufacturing mode, applying an SVD with arbitrary variables, rebooting
to the EFI shell, using dmpstore to dump the variables to a file,
applying that file in Config Editor and seeing that the current variable
state matches what was provided in the SVD.

## Integration Instructions

N/A