Updated dump option to put all variables into SVD format (#81)
## Description

This change is introduced to incorporate the XML knobs to be displayed
in the conf app.

Specifically, the conf app will iterate through all the available
variables and convert them
into base64 encoded variable list format as SVD files. The output will
be importable from
ConfigEditor UI tool for end user to view current settings.

For each item, place an "x" in between `[` and `]` if true. Example:
`[x]`.
_(you can also check items in the GitHub UI)_

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This change is tested on QEMU virtual platform.

## Integration Instructions

N/A. This change is transparent to end users.