Fix CSV Loading (#87)
# Preface

Please ensure you have read the [contribution
docs](https://github.com/microsoft/mu/blob/master/CONTRIBUTING.md) prior
to submitting the pull request. In particular,
[pull request
guidelines](https://github.com/microsoft/mu/blob/master/CONTRIBUTING.md#pull-request-best-practices).

## Description

Loading CSV files was brittle, depending on the CSV filename being the
same as the XML filename. This change allows the CSV filename to be
loaded if the XML filename is a prefix of the CSV filename. The docs are
updated with examples.

This will be updated to allow any CSV filename when the GUID is added to
the CSV output and loading knobs will instead look for the unique
GUID/knob name combination.

For each item, place an "x" in between `[` and `]` if true. Example:
`[x]`.
_(you can also check items in the GitHub UI)_

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [x] Includes documentation?

## How This Was Tested

Tested with some sample XML/CSV files.

## Integration Instructions

Ensure that CSV files have the XML filename as a prefix.