Restrict one-byte aligned packing to schema defined elements
Extending byte-aligned tight structure packing to types not defined in
the xml schema can lead to misaligned globals on platforms that have
strict alignment and packing requirements.

For this reason, only user defined elements from the xml schema should
be packed tightly. The alignment for rest of the auto-generated code
will be left upto the compiler and platform-level build settings.