SetupDataPkg: ConfApp: Dump configuration from policy service
This change leverages the new interfaces from ConfigVariableListLib to
dump configuration status from policy service.

In addition, the change removed a restriction of applying settings when
is already in manufacturing mode (regardless of the ready to boot state
or not), to match the new concept of security boundary.

The platform drivers are responsible for populating/setting the config
policy during earlier boot phase. The serial output from ConfApp will
remain the same SVD format as before.

Co-authored-by: Oliver Smith-Denny <osde@microsoft.com>
Signed-off-by: Kun Qin <kuqin@microsoft.com>